+++ example-projects/project A/.vscode/settings.json
@@ -18,8 +18,6 @@
         "@fast-skip-me",
         "@fast-skip-me-too"
     ],
-    // "behave-vsc.fastSkipList": "@fast-skip-me, @fast-skip-me-too",
-    // "behave-vsc.envVarList": "  'some_var':'double qu\"oted',  'some_var2':'single qu\\'oted', 'USERNAME': 'bob-163487', 'space_var': ' ' ",
     "behave-vsc.featuresPath": "behave tests/some tests/",
     "behave-vsc.justMyCode": true,
     "behave-vsc.runAllAsOne": true,
