+++ src/handlers/findStepReferencesHandler.ts
@@ -2,9 +2,9 @@ import * as vscode from 'vscode';
 import { config } from "../configuration";
 import { uriMatchString, getWorkspaceUriForFile, isStepsFile, openDocumentRange } from '../common';
 import { StepReference as StepReference, StepReferencesTree as StepReferencesTree } from './stepReferencesView';
-import { getStepMappingsForStepsFileFunction, waitOnReadyForStepsNavigation as refreshStepsNavigation } from '../parsing/stepMappings';
-import { FeatureFileStep } from '../parsing/featureParser';
-import { funcRe } from '../parsing/stepsParser';
+import { getStepMappingsForStepsFileFunction, waitOnReadyForStepsNavigation as refreshStepsNavigation } from '../parsers/stepMappings';
+import { FeatureFileStep } from '../parsers/featureParser';
+import { funcRe } from '../parsers/stepsParser';
