+++ src/handlers/gotoStepHandler.ts
@@ -1,8 +1,8 @@
 import * as vscode from 'vscode';
 import { config } from "../configuration";
 import { getWorkspaceUriForFile, isFeatureFile, openDocumentRange } from '../common';
-import { getStepFileStepForFeatureFileStep, waitOnReadyForStepsNavigation } from '../parsing/stepMappings';
-import { featureFileStepRe } from '../parsing/featureParser';
+import { getStepFileStepForFeatureFileStep, waitOnReadyForStepsNavigation } from '../parsers/stepMappings';
+import { featureFileStepRe } from '../parsers/featureParser';
