+++ src/runners/runOrDebug.ts
@@ -5,7 +5,7 @@ import { WorkspaceSettings } from "../settings";
 import { runAllAsOne, runScenario } from './behaveRun';
 import { debugScenario } from './behaveDebug';
 import { QueueItem } from '../extension';
-import { getJunitFileUri, updateTest } from '../parsing/junitParser';
+import { getJunitFileUri, updateTest } from '../parsers/junitParser';
 import { rndAlphaNumeric, WIN_MAX_PATH, WkspError } from '../common';
 import { cancelTestRun } from './testRunHandler';
